requirements.txt
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
pandas>=1.3.0
open3d>=0.15.0
laspy>=2.0.0
utm>=0.7.0
pyyaml>=6.0
dataclasses-json>=0.5.7
tqdm>=4.62.0
psutil>=5.8.0
plotly>=5.0.0
trimesh>=3.9.0
kaleido>=0.2.1